dump

+++ Math/binomial.cpp
@@ -1,16 +1,16 @@
 #include "binomial.h"
 
-std::vector<int> nth_combination(uint64_t n, uint64_t k, uint64_t index)
+std::vector<int8_t> nth_combination(int8_t n, int8_t k, int64_t index)
 {
-	std::vector<int> combination;
+	std::vector<int8_t> combination;
 	combination.reserve(k);
-	uint64_t remaining = k;
-	uint64_t start = 0;
+	int8_t remaining = k;
+	int8_t start = 0;
 	while (remaining > 0)
 	{
-		for (uint64_t i = start; i < n; ++i)
+		for (int8_t i = start; i < n; ++i)
 		{
-			uint64_t count = binomial(n - (i + 1), remaining - 1);
+			int64_t count = binomial(n - (i + 1), remaining - 1);
 			if (count > index)
 			{
 				combination.push_back(i);
